+++ src/update/main.cpp
@@ -27,6 +27,7 @@
 #include "Update_0.5.7_to_0.6.0.hpp"
 #include "Update_0.6.5_to_0.6.6.hpp"
 #include "UpdateTo_0.7.0.hpp"
+#include "UpdateTo_0.8.1.hpp"
 
 void showVersion()
 {
@@ -92,7 +93,7 @@ int main(int argc, char** argv)
         else
         {
           std::cerr << "Error: You have to enter a file path after \""
-                    << param <<"\"." << std::endl;
+                    << param << "\"." << std::endl;
           return wic::rcInvalidParameter;
         }
       } // if configuration file
@@ -134,6 +135,12 @@ int main(int argc, char** argv)
   }
   std::cout << "Update for database of version 0.6.7 (and earlier) to version 0.7.0..." << std::endl;
   if (!wic::UpdateTo070::perform(config.connectionInfo()))
+  {
+    std::cerr << "Error: Database update failed!" << std::endl;
+    return wic::rcUpdateFailure;
+  }
+  std::cout << "Update for database of version 0.8.0 (and earlier) to version 0.8.1..." << std::endl;
+  if (!wic::UpdateTo081::perform(config.connectionInfo()))
   {
     std::cerr << "Error: Database update failed!" << std::endl;
     return wic::rcUpdateFailure;